+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Changelog for aim.models
 2.1.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+### Added
+
+- sdb_cache field for Lambda.
 
 
 2.0.0 (2019-07-23)